requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=1.0
requests>=2.0
pandas>=2.0
azure-identity>=1.16.0
azure-storage-blob>=12.0.0
azure-communication-identity>=1.0.0
